fix: update service task rest to use /admin endpoint
