+++ activiti-cloud-query-service/activiti-cloud-services-query/activiti-cloud-services-query-rest/src/main/java/org/activiti/cloud/conf/QueryRestControllersAutoConfiguration.java
@@ -24,13 +24,13 @@ import org.activiti.cloud.services.query.rest.ProcessInstanceController;
 import org.activiti.cloud.services.query.rest.ProcessInstanceDeleteController;
 import org.activiti.cloud.services.query.rest.ProcessInstanceDiagramAdminController;
 import org.activiti.cloud.services.query.rest.ProcessInstanceDiagramController;
-import org.activiti.cloud.services.query.rest.ProcessInstanceServiceTasksController;
+import org.activiti.cloud.services.query.rest.ProcessInstanceServiceTasksAdminController;
 import org.activiti.cloud.services.query.rest.ProcessInstanceTasksController;
 import org.activiti.cloud.services.query.rest.ProcessInstanceVariableAdminController;
 import org.activiti.cloud.services.query.rest.ProcessInstanceVariableController;
 import org.activiti.cloud.services.query.rest.ProcessModelAdminController;
 import org.activiti.cloud.services.query.rest.ProcessModelController;
-import org.activiti.cloud.services.query.rest.ServiceTaskController;
+import org.activiti.cloud.services.query.rest.ServiceTaskAdminController;
 import org.activiti.cloud.services.query.rest.TaskAdminController;
 import org.activiti.cloud.services.query.rest.TaskController;
 import org.activiti.cloud.services.query.rest.TaskDeleteController;
@@ -63,8 +63,8 @@ import org.springframework.context.annotation.Import;
     TaskDeleteController.class,
     TaskVariableAdminController.class,
     TaskVariableController.class,
-    ServiceTaskController.class,
-    ProcessInstanceServiceTasksController.class
+    ServiceTaskAdminController.class,
+    ProcessInstanceServiceTasksAdminController.class
 })
 public class QueryRestControllersAutoConfiguration {
